/gopher/head: changeset 39
Fixed filesize misreporting bug in send_binary (where it would report
the size of foo.txt.gz to the client, then run it through a decoder
and send foo.txt which is much bigger). Error checking on rfopen()
calls, and logging for error condition failed writes to the client and
reads from disk.