patch: gopher 3.0.11 Makefile.config.in
Hello Mr. Goerzen,

I apologize if you are not the correct person to contact with this
matter.  If you happen to know who is maintaining this software I
would appreciate your passing this message to him/her.

Below is a patch against gopher 3.0.11 which allows "make install" to
honor DESTDIR when installing files bound for $(PREFIX)/bin and
$(PREFIX)/man/.../* .

Sincerely,

Tom Canich